+++ packages/server/src/create-instance/construct-style-tags.js
@@ -1,6 +1,6 @@
 // @flow
 import type { EmotionCache } from '@emotion/utils'
-import { generateStyleTag } from './inline'
+import { generateStyleTag } from './utils'
 
 const createConstructStyleTags = (
   cache: EmotionCache,
@@ -9,12 +9,18 @@ const createConstructStyleTags = (
   html: string,
   styles: Array<{ ids: Array<string>, css: string }>
 }) => {
-  return criticalData.styles.reduce((acc, item) => {
-    return (
-      acc +
-      generateStyleTag(cache.key, item.ids.join(' '), item.css, nonceString)
+  let styleTagsString = ''
+
+  criticalData.styles.forEach(item => {
+    styleTagsString += generateStyleTag(
+      cache.key,
+      item.ids.join(' '),
+      item.css,
+      nonceString
     )
-  }, '')
+  })
+
+  return styleTagsString
 }
 
 export default createConstructStyleTags
