+++ packages/server/src/create-instance/utils.js
@@ -1,0 +1,12 @@
+// @flow
+
+export function generateStyleTag(
+  cssKey: string,
+  ids: string,
+  styles: string,
+  nonceString: string
+) {
+  return `<style data-emotion="${cssKey} ${ids.substring(
+    1
+  )}"${nonceString}>${styles}</style>`
+}
